backend/agentcore/requirements.txt
aws-sdk-bedrock-runtime>=0.3.0,
prompt-toolkit>=3.0.52,
pyaudio>=0.2.14
strands-agents>=1.26.0

boto3>=1.40.0
botocore>=1.40.0